+++ packages/expo-router/src/__tests__/push.test.ios.tsx
@@ -59,17 +59,10 @@ it('stacks should always push a new route', () => {
               name: '(group)',
               params: {
                 id: '1',
-                screen: 'post/[id]',
-                params: {
-                  id: '1',
-                  params: {
-                    id: '1',
-                  },
-                  screen: 'index',
-                },
               },
               path: undefined,
               state: {
+                __internal__routerActionState: true,
                 index: 3,
                 key: expect.any(String),
                 routeNames: ['user/[id]', 'post/[id]'],
@@ -79,12 +72,9 @@ it('stacks should always push a new route', () => {
                     name: 'post/[id]',
                     params: {
                       id: '1',
-                      params: {
-                        id: '1',
-                      },
-                      screen: 'index',
                     },
                     state: {
+                      __internal__routerActionState: true,
                       index: 0,
                       key: expect.any(String),
                       routeNames: ['index'],
@@ -95,6 +85,7 @@ it('stacks should always push a new route', () => {
                           params: {
                             id: '1',
                           },
+                          path: '/post/1',
                         },
                       ],
                       stale: false,
@@ -105,13 +96,10 @@ it('stacks should always push a new route', () => {
                     name: 'user/[id]',
                     params: {
                       id: '1',
-                      params: {
-                        id: '1',
-                      },
-                      screen: 'index',
                     },
                     path: undefined,
                     state: {
+                      __internal__routerActionState: true,
                       index: 0,
                       key: expect.any(String),
                       routeNames: ['index'],
@@ -122,6 +110,7 @@ it('stacks should always push a new route', () => {
                           params: {
                             id: '1',
                           },
+                          path: '/user/1',
                         },
                       ],
                       stale: false,
@@ -132,13 +121,10 @@ it('stacks should always push a new route', () => {
                     name: 'post/[id]',
                     params: {
                       id: '2',
-                      params: {
-                        id: '2',
-                      },
-                      screen: 'index',
                     },
                     path: undefined,
                     state: {
+                      __internal__routerActionState: true,
                       index: 0,
                       key: expect.any(String),
                       routeNames: ['index'],
@@ -149,6 +135,7 @@ it('stacks should always push a new route', () => {
                           params: {
                             id: '2',
                           },
+                          path: '/post/2',
                         },
                       ],
                       stale: false,
@@ -159,13 +146,10 @@ it('stacks should always push a new route', () => {
                     name: 'user/[id]',
                     params: {
                       id: '1',
-                      params: {
-                        id: '1',
-                      },
-                      screen: 'index',
                     },
                     path: undefined,
                     state: {
+                      __internal__routerActionState: true,
                       index: 1,
                       key: expect.any(String),
                       routeNames: ['index'],
@@ -176,6 +160,7 @@ it('stacks should always push a new route', () => {
                           params: {
                             id: '1',
                           },
+                          path: '/user/1',
                         },
                         {
                           key: expect.any(String),
@@ -318,12 +303,10 @@ it('works in a nested layout Stack->Tab->Stack', () => {
             {
               key: expect.any(String),
               name: '(tabs)',
-              params: {
-                params: {},
-                screen: 'a',
-              },
+              params: {},
               path: undefined,
               state: {
+                __internal__routerActionState: true,
                 history: [
                   {
                     key: expect.any(String),
@@ -342,6 +325,7 @@ it('works in a nested layout Stack->Tab->Stack', () => {
                     key: expect.any(String),
                     name: 'a',
                     params: {},
+                    path: '/a',
                   },
                   {
                     key: expect.any(String),
@@ -351,11 +335,9 @@ it('works in a nested layout Stack->Tab->Stack', () => {
                   {
                     key: expect.any(String),
                     name: 'c',
-                    params: {
-                      params: {},
-                      screen: 'one',
-                    },
+                    params: {},
                     state: {
+                      __internal__routerActionState: true,
                       index: 2,
                       key: expect.any(String),
                       routeNames: ['one', 'two'],
@@ -364,6 +346,7 @@ it('works in a nested layout Stack->Tab->Stack', () => {
                           key: expect.any(String),
                           name: 'one',
                           params: {},
+                          path: '/c/one',
                         },
                         {
                           key: expect.any(String),
@@ -465,12 +448,10 @@ it('targets the correct Stack when pushing to a nested layout', () => {
             {
               key: expect.any(String),
               name: 'one',
-              params: {
-                params: {},
-                screen: 'index',
-              },
+              params: {},
               path: undefined,
               state: {
+                __internal__routerActionState: true,
                 index: 2,
                 key: expect.any(String),
                 routeNames: ['index', 'two', 'page'],
@@ -479,6 +460,7 @@ it('targets the correct Stack when pushing to a nested layout', () => {
                     key: expect.any(String),
                     name: 'index',
                     params: {},
+                    path: '/one',
                   },
                   {
                     key: expect.any(String),
@@ -489,12 +471,10 @@ it('targets the correct Stack when pushing to a nested layout', () => {
                   {
                     key: expect.any(String),
                     name: 'two',
-                    params: {
-                      params: {},
-                      screen: 'index',
-                    },
+                    params: {},
                     path: undefined,
                     state: {
+                      __internal__routerActionState: true,
                       index: 1,
                       key: expect.any(String),
                       routeNames: ['index', 'page'],
@@ -503,6 +483,7 @@ it('targets the correct Stack when pushing to a nested layout', () => {
                           key: expect.any(String),
                           name: 'index',
                           params: {},
+                          path: '/one/two',
                         },
                         {
                           key: expect.any(String),
@@ -591,15 +572,10 @@ it('push should also add anchor routes', () => {
             {
               key: expect.any(String),
               name: '(group)',
-              params: {
-                initial: false,
-                params: {
-                  initial: false,
-                },
-                screen: 'orange',
-              },
+              params: {},
               path: undefined,
               state: {
+                __internal__routerActionState: true,
                 index: 1,
                 key: expect.any(String),
                 routeNames: ['apple', 'index', 'orange'],
@@ -611,7 +587,8 @@ it('push should also add anchor routes', () => {
                   {
                     key: expect.any(String),
                     name: 'orange',
-                    params: { initial: false },
+                    params: {},
+                    path: '/orange',
                   },
                 ],
                 stale: false,
@@ -833,7 +810,9 @@ describe('singular', () => {
 
     // Should push /apple and remove all previous instances of /apple
     act(() => {
-      return router.push('/apple', { dangerouslySingular: (_, params) => params.slug?.toString() });
+      return router.push('/apple', {
+        dangerouslySingular: (_, params) => params.slug?.toString(),
+      });
     });
 
     expect(screen).toHaveRouterState({