+++ packages/expo-router/src/getLinkingConfig.ts
@@ -9,7 +9,6 @@ import type { UrlObject } from './global-state/getRouteInfoFromState';
 import type { StoreRedirects } from './global-state/router-store';
 import { getInitialURL, getPathFromState, getStateFromPath, subscribe } from './link/linking';
 import type { LinkingOptions } from './react-navigation/native';
-import { getActionFromState } from './react-navigation/native';
 import type { NativeIntent, RequireContext } from './types';
 
 export function getNavigationConfig(
@@ -142,9 +141,6 @@ export function getLinkingConfig(
         }) ?? '/'
       );
     },
-    // Add all functions to ensure the types never need to fallback.
-    // This is a convenience for usage in the package.
-    getActionFromState,
   };
 }